Pixel color extraction and counting
Two methods were added to process an image, one to go through each pixel and extract the RGB values of each pixel, the other one is a helper function to determine the distance from the passed pixel to the 6 main colors of the app's algorithm.
Removed imageMainColors methods as it was no longer required, now the getRGBs method does all the work with its helper function.
Added the user and its color relation map as member variables of the class of the fragment.