ColorForm and camera implementation updates
Added 2 more colors to the color form to have a wider variety of colors to relate images to. Modified Orange and Red so they could be closer to other colors and represent a closer and more viable option.
Removed the previous photo capture implementation as a cleaner and more efficient one was found on the Android docs.
Added an additional camera permission to the manifest.