ColorFormActivity visual update
Added 2 new spinners to get a better hold of the main colors of any image. Swapped all the colors for 8 new equally distributed colors on the 3D RGB space. All these new colors were added as XML resources and are accessed through the resource within the Java files, this requires some extra lines of code but makes it easier to trace the original color and modify all the uses of it in a single place.